move vfs
