+++ src/010-pipeline/build.ts
@@ -14,8 +14,8 @@ import { packageCode } from "../packaging/packageCode";
 import { createABITypeRefFromTypeRef } from "../types/resolveABITypeRef";
 import { getContracts, getType } from "../types/resolveDescriptors";
 import { posixNormalize } from "../utils/filePath";
-import { createVirtualFileSystem } from "../vfs/createVirtualFileSystem";
-import { VirtualFileSystem } from "../vfs/VirtualFileSystem";
+import { createVirtualFileSystem } from "../020-vfs/createVirtualFileSystem";
+import { VirtualFileSystem } from "../020-vfs/VirtualFileSystem";
 import { compile } from "./compile";
 import { precompile } from "./precompile";
 import { getCompilerVersion } from "./version";
