+++ src/010-pipeline/precompile.ts
@@ -6,7 +6,7 @@ import { resolveStatements } from "../types/resolveStatements";
 import { resolveErrors } from "../types/resolveErrors";
 import { resolveSignatures } from "../types/resolveSignatures";
 import { resolveImports } from "../imports/resolveImports";
-import { VirtualFileSystem } from "../vfs/VirtualFileSystem";
+import { VirtualFileSystem } from "../020-vfs/VirtualFileSystem";
 import { AstModule, FactoryAst } from "../grammar/ast";
 import { Parser } from "../grammar";
